backend/pyproject.toml
[project]
name = "bio-backend"
version = "0.1.0"
description = "Backend for professional portfolio website"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "pyyaml",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
